#6 Create a Basic Login Page
*Create login page
*Updated Home View page
