+++ src/router/index.js
@@ -12,10 +12,12 @@ const router = createRouter({
     {
       path: '/about',
       name: 'about',
-      // route level code-splitting
-      // this generates a separate chunk (About.[hash].js) for this route
-      // which is lazy-loaded when the route is visited.
       component: () => import('../views/AboutView.vue'),
+    },
+    {
+      path: '/login', // 👈 your new route
+      name: 'login',
+      component: () => import('../views/LoginView.vue'), // assumes you create this file
     },
   ],
 })
